Improves rebuild_docker_image script

+++ tests/rebuild_docker_image.sh
@@ -1,8 +1,12 @@
 #!/usr/bin/env bash
 set -e
 
-echo "Copying vm image from old docker image"
-docker run -it --rm -v $PWD:/host bkchr/hole_punch_tester:latest sh -c "cp /vm_image/hole_punch_vm.qcow2 /host/"
+if [ -z ${USE_EXISTING_IMAGE+x} ]; then
+  echo "Copying vm image from old docker image"
+  docker run -it --rm -v $PWD:/host bkchr/hole_punch_tester:latest sh -c "cp /vm_image/hole_punch_vm.qcow2 /host/"
+else
+  echo "Using existing image"
+fi
 
 echo "Creating new docker image"
 docker build --no-cache -t bkchr/hole_punch_tester:latest .
